pyproject.toml
[tool.poetry]
name = "cacomo"
version = "0.1.0"
description = "1st Machine Learning Model Training"
authors = ["Your Name <you@example.com>"]
readme = "README.md"


[tool.poetry.dependencies]
python = ">=3.12,<3.13"
transformers = "^4.51.3"
tqdm = "^4.67.1"
matplotlib = "^3.10.1"
pyyaml = "^6.0.2"
peft = "^0.15.2"
datasets = "^3.5.1"
streamlit = "^1.45.0"
sentencepiece = "^0.2.0"
torch = "^2.7.0"
nltk = "^3.9.1"
spacy = "^3.8.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"